fix in sitemap extensions traversal parent
this is fixing a strange behaviour when you are using
a URL structure like fr/some-page/other-page. Because
the sitemap object will look for the exact page with the
path including fr/ but fr/ is removed in line 18,
the page is not found. So we don't remove the fr/part.
This leads to a other problem. If using current_page.parent
and current_page.parent.parent it can be, that both these
are the same. So when creating a breadcrumb, you have to
check for that.